@@ -19,46 +19,37 @@
   ],
   "dependencies": {
     "classlist-polyfill": "^1.2.0",
-    "classnames": "^2.2.6",
-    "ease-component": "^1.0.0",
     "events-polyfill": "^2.1.2",
-    "fbjs": "^1.0.0",
-    "history": "^4.9.0",
-    "immutability-helper": "^3.0.1",
-    "lodash": "^4.17.14",
-    "mobile-detect": "^1.4.3",
+    "history": "^4.10.1",
+    "immer": "^4.0.1",
+    "mobile-detect": "^1.4.4",
     "modernizr": "^3.7.1",
+    "nanoid": "^2.1.2",
     "polished": "^3.4.1",
     "prop-types": "^15.7.2",
-    "qs": "^6.7.0",
-    "react": "^16.8.6",
-    "react-app-polyfill": "1.0.1",
-    "react-dom": "^16.8.6",
-    "react-helmet": "^5.2.1",
-    "react-inlinesvg": "^0.8.4",
-    "react-redux": "^7.1.0",
-    "react-router-dom": "^5.0.1",
-    "react-transition-group": "^4.2.1",
+    "qs": "^6.9.0",
+    "react": "^16.10.2",
+    "react-app-polyfill": "1.0.4",
+    "react-dom": "^16.10.2",
+    "react-helmet-async": "^1.0.3",
+    "react-inlinesvg": "^1.1.7",
+    "react-redux": "^7.1.1",
+    "react-router-dom": "^5.1.2",
+    "react-transition-group": "^4.3.0",
     "redux": "^4.0.4",
     "redux-actions": "^2.6.5",
-    "redux-immutable-state-invariant": "^2.1.0",
-    "redux-logger": "^3.0.6",
-    "redux-persist": "^5.10.0",
-    "redux-saga": "^1.0.5",
-    "sass-mediaqueries": "^1.6.1",
-    "scroll": "^3.0.1",
-    "scroll-doc": "^0.2.1",
-    "styled-components": "^4.3.2",
+    "redux-persist": "^6.0.0",
+    "redux-saga": "^1.1.1",
+    "styled-components": "^4.4.0",
     "styled-minimal": "^0.9.0",
-    "tree-changes": "^0.4.0",
-    "uuid": "^3.3.2"
+    "tree-changes": "^0.5.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.5.4",
+    "@babel/core": "^7.6.2",
     "@babel/plugin-external-helpers": "^7.2.0",
-    "@babel/plugin-proposal-class-properties": "7.5.0",
-    "@babel/plugin-proposal-decorators": "7.4.4",
-    "@babel/plugin-proposal-do-expressions": "7.5.0",
+    "@babel/plugin-proposal-class-properties": "7.5.5",
+    "@babel/plugin-proposal-decorators": "7.6.0",
+    "@babel/plugin-proposal-do-expressions": "7.6.0",
     "@babel/plugin-proposal-export-default-from": "7.5.2",
     "@babel/plugin-proposal-export-namespace-from": "7.5.2",
     "@babel/plugin-proposal-function-bind": "7.2.0",
@@ -67,120 +58,120 @@
     "@babel/plugin-proposal-logical-assignment-operators": "7.2.0",
     "@babel/plugin-proposal-nullish-coalescing-operator": "7.4.4",
     "@babel/plugin-proposal-numeric-separator": "7.2.0",
-    "@babel/plugin-proposal-optional-chaining": "7.2.0",
+    "@babel/plugin-proposal-optional-chaining": "7.6.0",
     "@babel/plugin-proposal-pipeline-operator": "7.5.0",
     "@babel/plugin-proposal-throw-expressions": "7.2.0",
     "@babel/plugin-syntax-dynamic-import": "7.2.0",
     "@babel/plugin-syntax-import-meta": "7.2.0",
     "@babel/plugin-transform-flow-strip-types": "^7.4.4",
     "@babel/plugin-transform-object-assign": "^7.2.0",
-    "@babel/plugin-transform-runtime": "^7.5.0",
-    "@babel/preset-env": "^7.5.4",
+    "@babel/plugin-transform-runtime": "^7.6.2",
+    "@babel/preset-env": "^7.6.2",
     "@babel/preset-flow": "^7.0.0",
     "@babel/preset-react": "^7.0.0",
-    "@babel/runtime": "^7.5.4",
-    "@hot-loader/react-dom": "^16.8.6",
-    "@svgr/webpack": "^4.3.1",
-    "@testing-library/cypress": "^4.0.4",
+    "@babel/runtime": "^7.6.2",
+    "@hot-loader/react-dom": "^16.9.0",
+    "@size-limit/preset-app": "^2.1.6",
+    "@svgr/webpack": "^4.3.3",
+    "@testing-library/cypress": "^5.0.1",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-eslint": "^10.0.2",
-    "babel-jest": "^24.8.0",
+    "babel-eslint": "^10.0.3",
+    "babel-jest": "^24.9.0",
     "babel-loader": "^8.0.6",
     "babel-plugin-array-includes": "^2.0.3",
     "babel-plugin-dynamic-import-node": "^2.3.0",
-    "babel-plugin-lodash": "^3.3.4",
-    "babel-plugin-named-asset-import": "^0.3.2",
+    "babel-plugin-named-asset-import": "^0.3.4",
     "babel-plugin-styled-components": "^1.10.6",
     "babel-plugin-transform-flow-strip-types": "^6.22.0",
-    "babel-plugin-transform-imports": "^2.0.0",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
     "bfj": "^7.0.1",
-    "bundlesize": "^0.18.0",
     "case-sensitive-paths-webpack-plugin": "^2.2.0",
     "chalk": "^2.4.2",
-    "circular-dependency-plugin": "^5.0.2",
-    "core-js": "^3.1.4",
-    "css-loader": "^3.0.0",
-    "cypress": "^3.4.0",
-    "date-fns": "^1.30.1",
-    "dotenv": "^8.0.0",
+    "circular-dependency-plugin": "^5.2.0",
+    "core-js": "^3.2.1",
+    "css-loader": "^3.2.0",
+    "cypress": "^3.4.1",
+    "date-fns": "^2.4.1",
+    "dotenv": "^8.1.0",
     "enzyme": "^3.10.0",
     "enzyme-adapter-react-16": "^1.14.0",
-    "enzyme-to-json": "^3.3.5",
-    "eslint": "^5.16.0",
-    "eslint-config-airbnb": "^17.1.1",
+    "enzyme-to-json": "^3.4.2",
+    "eslint": "^6.5.1",
+    "eslint-config-airbnb": "^18.0.1",
     "eslint-config-cypress": "^0.28.0",
-    "eslint-config-prettier": "^5.1.0",
+    "eslint-config-prettier": "^6.3.0",
+    "eslint-import-resolver-node": "^0.3.2",
     "eslint-import-resolver-webpack": "^0.11.1",
-    "eslint-loader": "^2.2.1",
+    "eslint-loader": "^3.0.2",
     "eslint-plugin-babel": "^5.3.0",
-    "eslint-plugin-flowtype": "^3.11.1",
-    "eslint-plugin-import": "^2.18.0",
+    "eslint-plugin-flowtype": "^4.3.0",
+    "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-prettier": "^3.1.0",
-    "eslint-plugin-react": "^7.14.2",
-    "eslint-plugin-redux-saga": "^1.0.0",
+    "eslint-plugin-prettier": "^3.1.1",
+    "eslint-plugin-react": "^7.16.0",
+    "eslint-plugin-react-hooks": "^2.1.2",
+    "eslint-plugin-redux-saga": "^1.1.1",
     "expose-loader": "^0.7.5",
-    "file-loader": "^4.0.0",
-    "flow-bin": "^0.102.0",
+    "file-loader": "^4.2.0",
+    "flow-bin": "^0.109.0",
     "fs-extra": "^8.1.0",
-    "git-info-webpack-plugin": "^1.0.1",
+    "git-rev-webpack-plugin": "^0.2.2",
     "hoist-non-react-statics": "^3.3.0",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^4.0.0-beta.7",
-    "husky": "^3.0.0",
-    "image-webpack-loader": "^5.0.0",
+    "husky": "^3.0.8",
+    "image-webpack-loader": "^6.0.0",
     "is-ci-cli": "^1.1.1",
-    "jest": "^24.8.0",
+    "jest": "^24.9.0",
     "jest-chain": "^1.1.2",
+    "jest-date-mock": "^1.0.7",
     "jest-environment-jsdom-global": "^1.2.0",
-    "jest-enzyme": "^7.0.2",
+    "jest-enzyme": "^7.1.1",
     "jest-extended": "^0.11.2",
     "jest-fetch-mock": "^2.1.2",
     "jest-localstorage-mock": "^2.4.0",
     "jest-serializer-html": "^7.0.0",
     "jest-styled-components": "^6.3.3",
-    "jest-watch-typeahead": "^0.3.1",
+    "jest-watch-typeahead": "^0.4.0",
     "json-loader": "^0.5.7",
-    "lint-staged": "^9.2.0",
-    "lodash-webpack-plugin": "^0.11.5",
-    "markdown-loader": "^5.0.0",
-    "mini-css-extract-plugin": "^0.7.0",
+    "lint-staged": "^9.4.1",
+    "markdown-loader": "^5.1.0",
+    "mini-css-extract-plugin": "^0.8.0",
     "modernizr-loader": "^1.0.1",
     "node-sass": "^4.12.0",
-    "offline-plugin": "^5.0.7",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
     "postcss-flexbugs-fixes": "^4.1.0",
     "postcss-loader": "^3.0.0",
     "postcss-preset-env": "^6.7.0",
     "postcss-safe-parser": "^4.0.1",
     "prettier": "^1.18.2",
-    "raw-loader": "^3.0.0",
-    "react-dev-utils": "^9.0.1",
-    "react-error-overlay": "^5.1.6",
-    "react-hot-loader": "^4.12.6",
+    "raw-loader": "^3.1.0",
+    "react-dev-utils": "^9.1.0",
+    "react-error-overlay": "^6.0.3",
+    "react-hot-loader": "^4.12.14",
+    "redux-immutable-state-invariant": "^2.1.0",
+    "redux-logger": "^3.0.6",
     "redux-mock-store": "^1.5.3",
     "redux-saga-test-plan": "^4.0.0-beta.3",
-    "regenerator-runtime": "^0.13.2",
+    "regenerator-runtime": "^0.13.3",
+    "repo-tools": "^0.2.0",
     "rsync": "^0.6.1",
-    "sass-loader": "^7.1.0",
-    "script-loader": "^0.7.2",
-    "serve": "^11.1.0",
-    "source-map-support": "^0.5.12",
-    "start-server-and-test": "^1.9.1",
-    "style-loader": "^0.23.1",
-    "stylelint": "^10.1.0",
-    "stylelint-config-standard": "^18.3.0",
+    "sass-loader": "^8.0.0",
+    "serve": "^11.2.0",
+    "source-map-support": "^0.5.13",
+    "start-server-and-test": "^1.10.4",
+    "style-loader": "^1.0.0",
+    "stylelint": "^11.0.0",
+    "stylelint-config-standard": "^19.0.0",
     "stylelint-config-styled-components": "^0.1.1",
-    "stylelint-order": "^3.0.1",
+    "stylelint-order": "^3.1.1",
     "stylelint-processor-styled-components": "^1.8.0",
-    "terser-webpack-plugin": "^1.3.0",
-    "url-loader": "^2.0.1",
-    "webpack": "^4.35.3",
-    "webpack-cli": "^3.3.5",
-    "webpack-dev-server": "^3.7.2",
-    "webpack-merge": "^4.2.1",
-    "yargs": "^13.2.4"
+    "terser-webpack-plugin": "^2.1.2",
+    "url-loader": "^2.2.0",
+    "webpack": "^4.41.0",
+    "webpack-cli": "^3.3.9",
+    "webpack-dev-server": "^3.8.2",
+    "workbox-webpack-plugin": "^4.3.1"
   },
   "scripts": {
     "start": "node tools/start",
