@@ -1,57 +1,33 @@
-// Polyfills
 import React from 'react';
 import ReactDOM from 'react-dom';
 
 import { Provider } from 'react-redux';
 import { PersistGate } from 'redux-persist/lib/integration/react';
+import { HelmetProvider } from 'react-helmet-async';
 
+import { showAlert } from 'actions';
 import { store, persistor } from 'store/index';
-import { showAlert } from 'actions/index';
 
-import App from 'containers/App';
 import Loader from 'components/Loader';
 import Reload from 'components/Reload';
 
-export const app = {
-  cssRetries: 0,
-  fetchRetries: 0,
+import App from './App';
+import { register } from './serviceWorker';
 
-  run() {
-    this.render(App);
+ReactDOM.render(
+  <Provider store={store}>
+    <PersistGate loading={<Loader size={100} block />} persistor={persistor}>
+      <HelmetProvider>
+        <App />
+      </HelmetProvider>
+    </PersistGate>
+  </Provider>,
+  document.getElementById('root'),
+);
 
-    /* istanbul ignore else */
-    if (process.env.NODE_ENV === 'production') {
-      this.initOfflinePlugin();
-    }
+/* istanbul ignore next */
+register({
+  onUpdate: () => {
+    store.dispatch(showAlert(<Reload />, { id: 'sw-update', icon: 'bolt', timeout: 0 }));
   },
-  initOfflinePlugin() {
-    const OfflinePlugin = require('offline-plugin/runtime');
-
-    /* istanbul ignore next */
-    OfflinePlugin.install({
-      onUpdateReady: () => {
-        OfflinePlugin.applyUpdate();
-      },
-      onUpdated: () => {
-        store.dispatch(showAlert(<Reload />, { id: 'sw-update', icon: 'bolt', timeout: 0 }));
-      },
-    });
-  },
-  render(Component) {
-    const root = document.getElementById('react');
-
-    /* istanbul ignore next */
-    if (root) {
-      ReactDOM.render(
-        <Provider store={store}>
-          <PersistGate loading={<Loader size={100} block />} persistor={persistor}>
-            <Component />
-          </PersistGate>
-        </Provider>,
-        root,
-      );
-    }
-  },
-};
-
-app.run();
+});
