@@ -1,6 +1,5 @@
-import { handleActions } from 'redux-actions';
-import immutable from 'immutability-helper';
 import { REHYDRATE } from 'redux-persist/lib/constants';
+import { handleActions } from 'modules/helpers';
 
 import { ActionTypes } from 'constants/index';
 
@@ -11,21 +10,15 @@ export const appState = {
 export default {
   app: handleActions(
     {
-      [REHYDRATE]: state =>
-        immutable(state, {
-          alerts: { $set: [] },
-        }),
-      [ActionTypes.HIDE_ALERT]: (state, { payload: { id } }) => {
-        const alerts = state.alerts.filter(d => d.id !== id);
-
-        return immutable(state, {
-          alerts: { $set: alerts },
-        });
+      [REHYDRATE]: draft => {
+        draft.alerts = [];
       },
-      [ActionTypes.SHOW_ALERT]: (state, { payload }) =>
-        immutable(state, {
-          alerts: { $push: [payload] },
-        }),
+      [ActionTypes.HIDE_ALERT]: (draft, { payload: { id } }) => {
+        draft.alerts = draft.alerts.filter(d => d.id !== id);
+      },
+      [ActionTypes.SHOW_ALERT]: (draft, { payload }) => {
+        draft.alerts.push(payload);
+      },
     },
     appState,
   ),
