@@ -1,6 +1,5 @@
-import { handleActions } from 'redux-actions';
-import immutable from 'immutability-helper';
 import { parseError } from 'modules/client';
+import { handleActions } from 'modules/helpers';
 
 import { ActionTypes, STATUS } from 'constants/index';
 
@@ -16,36 +15,22 @@ export const githubState = {
 export default {
   github: handleActions(
     {
-      [ActionTypes.GITHUB_GET_REPOS]: (state, { payload }) => {
-        const data = state.repos.data[payload.query] ? state.repos.data[payload.query] : [];
-
-        return immutable(state, {
-          repos: {
-            data: {
-              [payload.query]: { $set: data },
-            },
-            message: { $set: '' },
-            query: { $set: payload.query },
-            status: { $set: STATUS.RUNNING },
-          },
-        });
+      [ActionTypes.GITHUB_GET_REPOS]: (draft, { payload }) => {
+        draft.repos.data[payload.query] = draft.repos.data[payload.query]
+          ? draft.repos.data[payload.query]
+          : [];
+        draft.repos.message = '';
+        draft.repos.query = payload.query;
+        draft.repos.status = STATUS.RUNNING;
       },
-      [ActionTypes.GITHUB_GET_REPOS_SUCCESS]: (state, { payload }) =>
-        immutable(state, {
-          repos: {
-            data: {
-              [state.repos.query]: { $set: payload.data || [] },
-            },
-            status: { $set: STATUS.READY },
-          },
-        }),
-      [ActionTypes.GITHUB_GET_REPOS_FAILURE]: (state, { payload }) =>
-        immutable(state, {
-          repos: {
-            message: { $set: parseError(payload.message) },
-            status: { $set: STATUS.ERROR },
-          },
-        }),
+      [ActionTypes.GITHUB_GET_REPOS_SUCCESS]: (draft, { payload }) => {
+        draft.repos.data[draft.repos.query] = payload.data || [];
+        draft.repos.status = STATUS.SUCCESS;
+      },
+      [ActionTypes.GITHUB_GET_REPOS_FAILURE]: (draft, { payload }) => {
+        draft.repos.message = parseError(payload.message);
+        draft.repos.status = STATUS.ERROR;
+      },
     },
     githubState,
   ),
