@@ -1,5 +1,4 @@
-import { handleActions } from 'redux-actions';
-import immutable from 'immutability-helper';
+import { handleActions } from 'modules/helpers';
 
 import { STATUS, ActionTypes } from 'constants/index';
 
@@ -11,24 +10,20 @@ export const userState = {
 export default {
   user: handleActions(
     {
-      [ActionTypes.USER_LOGIN]: state =>
-        immutable(state, {
-          status: { $set: STATUS.RUNNING },
-        }),
-      [ActionTypes.USER_LOGIN_SUCCESS]: state =>
-        immutable(state, {
-          isAuthenticated: { $set: true },
-          status: { $set: STATUS.READY },
-        }),
-      [ActionTypes.USER_LOGOUT]: state =>
-        immutable(state, {
-          status: { $set: STATUS.RUNNING },
-        }),
-      [ActionTypes.USER_LOGOUT_SUCCESS]: state =>
-        immutable(state, {
-          isAuthenticated: { $set: false },
-          status: { $set: STATUS.IDLE },
-        }),
+      [ActionTypes.USER_LOGIN]: draft => {
+        draft.status = STATUS.RUNNING;
+      },
+      [ActionTypes.USER_LOGIN_SUCCESS]: draft => {
+        draft.isAuthenticated = true;
+        draft.status = STATUS.READY;
+      },
+      [ActionTypes.USER_LOGOUT]: draft => {
+        draft.status = STATUS.RUNNING;
+      },
+      [ActionTypes.USER_LOGOUT_SUCCESS]: draft => {
+        draft.isAuthenticated = false;
+        draft.status = STATUS.IDLE;
+      },
     },
     userState,
   ),
