@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { App } from 'containers/App';
+import { App } from 'App';
 
 const mockDispatch = jest.fn();
 
@@ -22,10 +22,7 @@ describe('App', () => {
   const wrapper = setup();
 
   it('should render properly for anonymous users', () => {
-    expect(wrapper.find('HelmetWrapper')).toExist();
-    expect(wrapper.find('Switch')).toExist();
-    expect(wrapper.find('Footer')).toExist();
-    expect(wrapper.find('Connect(SystemAlerts)')).toExist();
+    expect(wrapper.debug()).toMatchSnapshot();
   });
 
   it('should render properly for logged users', () => {
